Fix TS 5.9 Uint8Array generic incompatibility in specter-sdk crypto
Cast sha256() result to Uint8Array to fix ArrayBufferLike vs ArrayBuffer

+++ packages/specter-sdk/src/utils/crypto.ts
@@ -173,7 +173,7 @@ export function deriveKeyFromPassword(
 
   // Simple iterative hashing (use proper PBKDF2 in production)
   for (let i = 0; i < Math.min(KDF_ITERATIONS, 10000); i++) {
-    key = sha256(key);
+    key = sha256(key) as Uint8Array;
   }
 
   return key;
